Fix baseurl issue https://github.com/poole/poole/issues/10
Adding a preceding slash makes the browser look for the assets in the root, whereas it should be looking one level down in the repo directory. In order to make it work, I added a baseurl param in _config.yml and used it at the assets load.
To preview the site locally use the --baseurl flag, but without a value: jekyll serve --baseurl=